function/requirements.txt
azure-functions
azure-storage-blob
azure-identity
pydantic>=2.0
